+++ src/modules/module_22001.c
@@ -920,17 +920,17 @@ int module_hash_decode (MAYBE_UNUSED const hashconfig_t *hashconfig, MAYBE_UNUSE
 
     wpa->pmkid_data[0] = 0x204b4d50; // "PMK "
     wpa->pmkid_data[1] = 0x656d614e; // "Name"
-    wpa->pmkid_data[2] = (mac_ap[0]  <<  0)
-                       | (mac_ap[1]  <<  8)
-                       | (mac_ap[2]  << 16)
-                       | ((u32) mac_ap[3] << 24);
-    wpa->pmkid_data[3] = (mac_ap[4]  <<  0)
-                       | (mac_ap[5]  <<  8)
-                       | (mac_sta[0] << 16)
+    wpa->pmkid_data[2] = ((u32) mac_ap[0]  <<  0)
+                       | ((u32) mac_ap[1]  <<  8)
+                       | ((u32) mac_ap[2]  << 16)
+                       | ((u32) mac_ap[3]  << 24);
+    wpa->pmkid_data[3] = ((u32) mac_ap[4]  <<  0)
+                       | ((u32) mac_ap[5]  <<  8)
+                       | ((u32) mac_sta[0] << 16)
                        | ((u32) mac_sta[1] << 24);
-    wpa->pmkid_data[4] = (mac_sta[2] <<  0)
-                       | (mac_sta[3] <<  8)
-                       | (mac_sta[4] << 16)
+    wpa->pmkid_data[4] = ((u32) mac_sta[2] <<  0)
+                       | ((u32) mac_sta[3] <<  8)
+                       | ((u32) mac_sta[4] << 16)
                        | ((u32) mac_sta[5] << 24);
 
     // hash
